+++ openmdao/test_suite/components/metadata_feature_lincomb.py
@@ -1,0 +1,21 @@
+"""
+Component for a metadata feature test.
+"""
+import numpy as np
+
+from openmdao.api import ExplicitComponent
+
+
+class LinearCombinationComp(ExplicitComponent):
+
+    def initialize(self):
+        self.metadata.declare('a', default=1., type_=np.ScalarType)
+        self.metadata.declare('b', default=1., type_=np.ScalarType)
+
+    def setup(self):
+        self.add_input('x')
+        self.add_output('y')
+        self.declare_partials('y', 'x', val=self.metadata['a'])
+
+    def compute(self, inputs, outputs):
+        outputs['y'] = self.metadata['a'] * inputs['x'] + self.metadata['b']
